Return JSON from jev_eval and drop C-style casts
jev_eval now casts the answer text to JSON, as docs/DESIGN.md specifies. The
printf-style casts around StringUtil::Format arguments are gone - fmt reads the
argument types itself, and clang-tidy rejects C-style casts.

+++ src/jev_client.cpp
@@ -263,7 +263,7 @@ JevResponse JevPostBatch(const JevConfig &config, const string &question, const 
 		}
 		last_error = StringUtil::Format("TypeSafe API error %d %s", res->status, Truncate(res->body, 300));
 	}
-	throw InvalidInputException("jev: %s (gave up after %llu attempts)", last_error, (unsigned long long)MAX_ATTEMPTS);
+	throw InvalidInputException("jev: %s (gave up after %llu attempts)", last_error, MAX_ATTEMPTS);
 }
 
 } // namespace duckdb
